+++ storage/perfschema/pfs_global.cc
@@ -18,6 +18,7 @@
   Miscellaneous global dependencies (implementation).
 */
 
+#include <my_global.h>
 #include "pfs_global.h"
 #include <my_sys.h>
 #include <my_net.h>
@@ -218,6 +219,6 @@ uint pfs_get_socket_address(char *host,
   }
 
   /* Return actual IP address string length */
-  return (strlen((const char*)host));
+  return ((uint)strlen((const char*)host));
 }
